fix: goroutine leak in Client.resetActivityTimer method
This PR fixes a goroutine leak in `Client.resetActivityTimer` func.

The use case for `Client` is to be first instantiated (e.g. via `Client{}`) and then to call `Client.Connect`. The `Connect` method populates additional internal fields of `Client` when connection is established successfully.
An instance of `Client` created this way always had `activityTimerReset` channel not initialized (`nil`). This led to always blocked writes and reads from this channel.

`Client.resetActivityTimer` tried to write to the `activityTimerReset` channel in goroutine. Because of the `nil` channel, the writes could not finish and the goroutine would not return creating a goroutine leak.

The reads from `activityTimerReset` done in `Client.heartbeat` method were also blocked. Most likely, that essentially disabled heartbeat exchange after first attempt.

Here's a quick example: https://go.dev/play/p/bFu8azDKlBl.

This PR makes sure the `activityTimerReset` is initialized by `Client.Connect` method. It also changes the channel to be buffered (buffer size 1) and removes a goroutine wrap.

The `activityTimerReset` channel is used to send a signal that the timer needs to be reset at a certain moment. We use a non-blocking write implemented with `select-default` to remove the need for a goroutine. The buffered channel of size 1 allows us to not loose reset signal when `Client.heartbeat` executes its second `case` statement – sending ping message to websocket. Once exchange finishes, `Client.heartbeat` will read from `activityTimerReset` and will reset the timer.